- Customizer cleanup part 1  - Arrow positioning cleanup  - Cursors!  Lots of cursors!  - EnhRectangle2D enhancements for use collecting repaint regions  - PathElement can apply itself to a Path2D  - Rewrite of PathTool  - Stacks in ContextLog  - Zoom now has a stock set of zoom levels and uses double precision  - Generic anno processor for random things tools choose from, replace layer code with that where possible  - Get rid of unused layers customizer  - More svg icons, eating our own dogfood  - Rewrite of TitledPanel to be sane and readable, with umpteen layout managers to manage complex indentic  - Use font-scaled borders  - SimpleSliderUI bugfixes / tracking rewrite  - Adjust colors in gradient customizers  - Waffle tool - create a pattern - an incomplete path, and have it interpolated between points you draw; with pluggable interpolaters

+++ Imagine/paintui/src/main/java/org/netbeans/paintui/PaintCanvas.java
@@ -167,9 +167,9 @@ class PaintCanvas extends JComponent implements RepaintHandle, ChangeListener, P
         }
         return result;
     }
-    private float zoom = 1.0f;
+    private double zoom = 1.0f;
 
-    public void setZoom(float zoom) {
+    public void setZoom(double zoom) {
         if (this.zoom != zoom) {
             Dimension old = getPreferredSize();
             this.zoom = zoom;
@@ -181,7 +181,7 @@ class PaintCanvas extends JComponent implements RepaintHandle, ChangeListener, P
         }
     }
 
-    public float getZoom() {
+    public double getZoom() {
         return zoom;
     }
     private Rectangle imageBounds = new Rectangle();
@@ -382,7 +382,7 @@ class PaintCanvas extends JComponent implements RepaintHandle, ChangeListener, P
         if ((e.getModifiersEx() & MouseWheelEvent.SHIFT_DOWN_MASK) != 0) {
             //Use Shift-MouseWheel to zoom
             float ct = -e.getWheelRotation();
-            float zoomFactor = getZoom();
+            double zoomFactor = getZoom();
             zoomFactor += ct / 10;
             zoomFactor = Math.max(0.1f, zoomFactor);
             zoomImpl.setZoom(zoomFactor);
@@ -606,11 +606,11 @@ class PaintCanvas extends JComponent implements RepaintHandle, ChangeListener, P
 
         private ChangeSupport supp = new ChangeSupport(this);
 
-        public float getZoom() {
+        public double getZoom() {
             return PaintCanvas.this.getZoom();
         }
 
-        public void setZoom(float val) {
+        public void setZoom(double val) {
             if (val != getZoom()) {
                 PaintCanvas.this.setZoom(val);
                 supp.fireChange();
